requirements.txt
sentence-transformers
faiss-cpu
torch
accelerate

fastapi
uvicorn
python-multipart

groq
python-dotenv
langdetect
pydantic
PyMuPDF